third_party/blink/web_tests/virtual/change-service-worker-priority-when-client-foreground-state-change/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634042632175.5 but got 1634042632172.9001
Harness: the test ran to completion.
